Added several lines of debug text, modified multijump logic
Modified several print statements to only occur if settings["debug"] is True
Modified multijump logic to prevent an infite loop.
Added stop_search bool to further_possible_moves loop, which is set to True if a
box to be added to the list is already present in the list.  False by default.