Fixed bug: Modified kill list behavior to allow longer list
Modified kill list to append new entries instead of overwriting the first entry.
Also added check to make sure the same box is not added multiple times to the list,
as lists briefly ballooned and took on far more data than was needed.
Previously, kill list was restricted to 1 entry, which meant that in 3+ multijumps,
manhy valid pieces were not removed when they should have been.